+++ src/Shipment/Contracts/ShippingMethod.php
@@ -15,8 +15,13 @@ declare(strict_types=1);
 namespace Vanilo\Shipment\Contracts;
 
 use Vanilo\Contracts\Configurable;
+use Vanilo\Shipment\Models\ShippingFee;
 
 interface ShippingMethod extends Configurable
 {
     public function getCarrier(): ?Carrier;
+
+    public function getCalculator(): ShippingFeeCalculator;
+
+    public function estimate(?object $subject = null): ShippingFee;
 }