+++ Changelog.md
@@ -113,6 +113,7 @@
     - `setPayableRemoteId()`
     - `findByPayableRemoteId()`
 - BC: The return type of the `getNumber()` method of the Order interface is no longer nullable
+- BC: Added the `getCalculator()` & `estimate()` methods to the `ShippingMethod` interface
 - BC: The `Channel` interface extends the `Configurable` interface
 - BC: Added the following methods to the `Channel` interface:
     - `getLanguage()`